Added init file and rename module
